Alterações para o teste. Os valores de A e B no ID das operações com float são mostrados como iguais aos esperados, mas os if não reconhece. Modificar alterar o tipo (com (int)) na hora da comparação parece resolver. Todos os outros testes acontecem sem problemas.

+++ fortnight03/runtest.cpp
@@ -452,7 +452,7 @@ void testID(BasicCPUTest* cpu,
 	long B = cpu->getB();
 	cout << "	A=0x" << A << "; B=0x" << B << endl;
 	cout << "Expected: A=0x" << xpctdA << "; B=0x" << xpctdB << endl;
-	if ((A != xpctdA) || (B != xpctdB)){
+	if (((int)A != xpctdA) || ((int)B != xpctdB)){
 		cout << "ID() FAILED on registers reading!" << endl;
 		cout << "Exit..." << endl;
 		exit(1);
